refactor: move NoUri to sphinx.errors
