+++ CHANGES
@@ -13,6 +13,7 @@ Incompatible changes
 Deprecated
 ----------
 
+* ``sphinx.environment.NoUri``
 * ``sphinx.ext.autodoc.importer.MockFinder``
 * ``sphinx.ext.autodoc.importer.MockLoader``
 * ``sphinx.ext.autodoc.importer.mock()``
@@ -20,6 +21,8 @@ Deprecated
 * ``sphinx.util.i18n.find_catalog()``
 * ``sphinx.util.i18n.find_catalog_files()``
 * ``sphinx.util.i18n.find_catalog_source_files()``
+
+For more details, see :ref:`deprecation APIs list <dev-deprecated-apis>`.
 
 Features added
 --------------
